backend/requirements.txt
fastapi
uvicorn
openai
openai-whisper
opencv-python
pytesseract
yt-dlp
torch
transformers